Add body iterator interface
